Live session onboarding
