+++ client/src/lib/daily/DailyProvider.tsx
@@ -19,6 +19,7 @@ export type DailyProviderTypes = {
   hasCameraPermissions: () => boolean;
   hasMicrophonePermissions: () => boolean;
   preJoinMeeting: (url: string, token: string) => Promise<void>;
+  startCamera: () => Promise<void>;
   joinMeeting: (options?: DailyCallOptions) => Promise<void>;
   leaveMeeting: () => Promise<void>;
   toggleAudio: (enabled: boolean) => void;
@@ -34,6 +35,7 @@ export const DailyContext = createContext<DailyProviderTypes>({
   hasCameraPermissions: () => false,
   hasMicrophonePermissions: () => false,
   preJoinMeeting: () => Promise.resolve(),
+  startCamera: () => Promise.resolve(),
   joinMeeting: () => Promise.resolve(),
   leaveMeeting: () => Promise.resolve(),
   toggleAudio: () => {},
@@ -119,19 +121,6 @@ const DailyProvider: React.FC<{children: React.ReactNode}> = ({children}) => {
     await daily.leave();
   }, [daily]);
 
-  const prepareMeeting = useCallback(
-    async (url: string, token: string) => {
-      if (daily.meetingState() !== 'joined-meeting') {
-        await daily.preAuth({
-          url,
-          token,
-        });
-      }
-    },
-
-    [daily],
-  );
-
   const setSubscribeToAllTracks = useCallback(() => {
     if (!daily) {
       return;
@@ -176,12 +165,21 @@ const DailyProvider: React.FC<{children: React.ReactNode}> = ({children}) => {
   const preJoinMeeting = useCallback(
     async (url: string, token: string) => {
       if (daily.meetingState() === 'new') {
-        await prepareMeeting(url, token);
-        await daily.startCamera({url});
+        await daily.preAuth({
+          url,
+          token,
+        });
+        //await daily.startCamera({url});
       }
     },
-    [daily, prepareMeeting],
+    [daily],
   );
+
+  const startCamera = useCallback(async () => {
+    if (daily.meetingState() !== 'joined-meeting') {
+      await daily.startCamera();
+    }
+  }, [daily]);
 
   const joinMeeting = useCallback(
     async (options?: DailyCallOptions) => {
@@ -252,6 +250,7 @@ const DailyProvider: React.FC<{children: React.ReactNode}> = ({children}) => {
         hasCameraPermissions,
         hasMicrophonePermissions,
         preJoinMeeting,
+        startCamera,
         joinMeeting,
         leaveMeeting,
         toggleAudio,
